MBP-131: Update advanced homing
Update pointer to tc_mca_std_lib

Fix reference to MC2 library: changed to 3.3.28.0 instead of 3.3.31.0

There may be some changes to the TcUnit tests but these were not
intended to be part of the commit. As I am halfway through the merge I
will commit what is there and then fix later on if required. These
should be part of another ticket.